[FTR][Saved Query Management] Split `saved query management/saved_query_management` config (#256715)
## Summary

Split
`x-pack/platform/test/functional/apps/saved_query_management/config.ts`
(~33 minutes) into smaller groups targeting 15–20 minutes each.

- Split `config.ts` into 2 groups by partitioning 4 feature security
tests (discover+dashboard in `group1/`, maps+visualize in `group2/`)
- Moved `config.ts` and `security.ts` via `git mv` to `group1/`; created
new `group2/` with its own `config.ts`, `index.ts`, and `security.ts`
- Registered both new configs in
`.buildkite/ftr_platform_stateful_configs.yml` replacing the original
entry

> [!NOTE]
> This PR was generated automatically by an AI coding agent ([Buildkite
build](https://buildkite.com/elastic/appex-qa-ai-ftr-config-splitter/builds/31)).
> Please review carefully before merging.

(cherry picked from commit 070318c9d996ba19ddc486b2fda9d0edc9290d3d)
