+++ x-pack/platform/test/functional/apps/saved_query_management/group2/feature_controls/v1/security.ts
@@ -1,0 +1,28 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License
+ * 2.0; you may not use this file except in compliance with the Elastic License
+ * 2.0.
+ */
+
+import type { FeatureApp, FeatureName } from '../../../feature_controls/create_security_tests';
+import { createSecurityTests } from '../../../feature_controls/create_security_tests';
+
+const featureConfigs: Array<{
+  feature: FeatureName;
+  app: FeatureApp;
+  hasImplicitSaveQueryManagement: boolean;
+}> = [
+  {
+    feature: 'maps',
+    app: 'maps',
+    hasImplicitSaveQueryManagement: true,
+  },
+  {
+    feature: 'visualize',
+    app: 'visualize',
+    hasImplicitSaveQueryManagement: true,
+  },
+];
+
+export default createSecurityTests(featureConfigs);
